# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ETF-Insight API
  description: ETF 数据管理与分析平台 API
  version: 2.6.0
  contact:
    name: ETF-Insight Team
    email: coder100001@gmail.com
servers:
  - url: http://localhost:8080
    description: 本地开发服务器
...
paths:
...
  /api/etf/list:
    get:
      tags:
        - ETF
      summary: 获取 ETF 列表
      description: 获取所有 ETF 列表，包含实时数据，支持分页
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
          description: 页码，从1开始
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
          description: 每页数量，最大100
      responses:
        '200':
          description: 成功返回 ETF 列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ETF'
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                        description: 当前页码
                      pageSize:
                        type: integer
                        description: 每页数量
                      total:
                        type: integer
                        description: 总记录数
                      totalPages:
                        type: integer
                        description: 总页数
...
components:
  schemas:
    ETF:
      type: object
      properties:
        symbol:
          type: string
          example: SCHD
        name:
          type: string
          example: Schwab US Dividend Equity ETF
        market:
          type: string
          example: US
        category:
          type: string
          example: 股息
        current_price:
          type: number
          format: float
          example: 30.44
        change:
          type: number
          format: float
          example: -1.23
        change_percent:
          type: number
          format: float
          example: -3.88
        dividend_yield:
          type: number
          format: float
          example: 3.45
...